+++ src/Arch/CSky/Registers.cs
@@ -83,8 +83,8 @@ namespace Reko.Arch.CSky
             Psr = CodesToControlRegisters[0];
             Vbr = CodesToControlRegisters[1];
             //$TODO: manual is unclear on the positions of the bits in the psr
-            C = new FlagGroupStorage(Psr, (uint) FlagM.CF, nameof(C));
-            V = new FlagGroupStorage(Psr, (uint) FlagM.VF, nameof(V));
+            C = new FlagGroupStorage(Psr, (ulong) FlagM.CF, nameof(C));
+            V = new FlagGroupStorage(Psr, (ulong) FlagM.VF, nameof(V));
         }
 
         public static RegisterStorage ControlRegister(uint ireg)
